requirements.txt
Augmentor
discord.py
numpy
aiohttp
pillow
pytz
platform
asyncio
pathlib
os
random
shutil
